requirements.txt
requests
keepa
flask==3.0.3
gunicorn==22.0.0
requests>=2.31
httpx>=0.27
langchain>=0.3,<0.4
langchain-core>=0.3.77,<1.0
langchain-openai>=0.3.33,<1.0
langchain-text-splitters>=0.3,<0.4
langgraph>=0.2
keepa>=1.3  
google-search-results>=2.4  # SerpApi official client
streamlit>=1.36
pydantic>=2.7
numpy>=1.26
pandas>=2.2
fastapi>=0.110
uvicorn[standard]>=0.30
Flask>=3.0
pytest>=8.2
pytest-cov>=5.0
python-dotenv
pathlib
sentence-transformers
faiss-cpu
psycopg[binary]>=3.2
sqlite3



